Separate the code of libraries from client's code
