+++ dictionary_api.cpp
@@ -1,5 +1,4 @@
-#include "yandextranslate.h"
-#include "ui_yandextranslate.h"
+#include "dictionary_api.h"
 
 #include <QNetworkAccessManager>
 #include <QNetworkRequest>
@@ -13,170 +12,102 @@
 #include <QJsonArray>
 #include <QJsonValue>
 
-#include <QVector>
+#include <QDebug>
 #include <QMessageBox>
-#include "dictionaryentry.h"
 
-YandexTranslate::YandexTranslate(QWidget *parent) :
-    QMainWindow(parent),
-    ui(new Ui::YandexTranslate)
+Example::Example()
+{}
+
+Example::~Example()
+{}
+
+Translation::Translation()
+{}
+
+Translation::~Translation()
+{}
+
+DictionaryEntry::DictionaryEntry()
+{}
+
+DictionaryEntry::~DictionaryEntry()
+{}
+
+YandexDictionary::YandexDictionary()
+{}
+
+YandexDictionary::~YandexDictionary()
+{}
+
+void Example::showExample()
 {
-    // Initial setup.
-    ui->setupUi(this);
-    this->setFixedSize(this->geometry().width(),this->geometry().height());
-    this->getSupportedLanguages();
-    this->getTranslationDirections();
+    qDebug() << this->text; // Text of an example.
 
-    ui->comboBoxSource->setCurrentText("English");
-    ui->comboBoxTarget->setCurrentText("Russian");
-
-    connect(ui->detectButton,SIGNAL(clicked(bool)),this,SLOT(detectLanguage()));
-    connect(ui->translateButton,SIGNAL(clicked(bool)),this,SLOT(translate()));
-    connect(ui->dictionaryButton,SIGNAL(clicked(bool)),this,SLOT(getDictionaryEntry()));
-}
-
-YandexTranslate::~YandexTranslate()
-{
-    delete ui;
-}
-
-void YandexTranslate::getSupportedLanguages()
-{
-    // Create a request.
-    QUrl url(QString("https://translate.yandex.net/api/v1.5/tr.json/getLangs?")
-             + QString("key=") + trnsl_api_key
-             + QString("&ui=") + QString("en")
-             );
-
-    QNetworkRequest request(url);
-    request.setHeader(QNetworkRequest::ContentTypeHeader, "application/x-www-form-urlencoded");
-
-    // Send the request.
-    QEventLoop loop;
-    QNetworkAccessManager manager;
-    connect(&manager, &QNetworkAccessManager::finished, &loop, &QEventLoop::quit);
-
-    // Receive a reply.
-    QNetworkReply* reply = manager.get(request);
-    loop.exec();
-
-    // Parse the reply.
-    QByteArray bytes = reply->readAll();
-    QJsonObject obj = QJsonDocument::fromJson(bytes).object();  // Keys are ("dirs", "langs")
-    // qDegub() << obj.keys();  // Use it to check the keys.
-
-    // Check the response codes.
-    if(obj.contains("code") && obj.contains("message"))
+    QStringList::const_iterator iter;
+    for(iter = this->translations.constBegin(); iter != this->translations.constEnd(); ++iter)
     {
-        QMessageBox::information(this,QString::number(obj.value("code").toDouble()),obj.value("message").toString());
-        return;
-    }
-
-    /* Use the following commented line to check the type of an object
-     * when you are not sure.
-     * There're also such functions as isArray(), isDouble(), etc.
-     * See Qt Documentation for more examples. */
-    //qDebug() << obj.value("langs").isObject();
-
-    QJsonObject langs = obj.value("langs").toObject();
-    for(const QString& code : langs.keys())
-    {
-        QString lang = langs.value(code).toString();    // "lang" is a full name of a language.
-        languages[code] = lang;
-        // Fill comboBoxes with initial data.
-        ui->comboBoxSource->addItem(lang);
-        ui->comboBoxTarget->addItem(lang);
+        qDebug() << *iter;  // Translations of the text of an example.
     }
 }
 
-void YandexTranslate::detectLanguage()
+void Translation::showTranslation()
 {
-    // Create a request.
-    QString source_text = ui->SourceEdit->toPlainText();
-    QUrl url(QString("https://translate.yandex.net/api/v1.5/tr.json/detect?")
-             + QString("key=") + trnsl_api_key
-             + QString("&text=") + source_text
-             //+ QString("&hint=") + <list of probable text languages>
-             );
+    QString header = (this->text + " "
+                      + this->pos + " "
+                      + this->gen + " "
+                      + this->asp + " "
+                      + this->num); // Header of a translation.
+    qDebug() << header.trimmed();   // Get rid of spaces in the beginning and in the end of a string.
 
-    QNetworkRequest request(url);
-    request.setHeader(QNetworkRequest::ContentTypeHeader, "application/x-www-form-urlencoded");
-
-    // Send the request.
-    QEventLoop loop;
-    QNetworkAccessManager manager;
-    connect(&manager, &QNetworkAccessManager::finished, &loop, &QEventLoop::quit);
-
-    // Receive a reply.
-    QNetworkReply* reply = manager.get(request);
-    loop.exec();
-
-    // Parse the reply.
-    QByteArray bytes = reply->readAll();
-    QJsonObject obj = QJsonDocument::fromJson(bytes).object();  // Keys are ("code", "lang")
-
-    // Check the response codes.
-    if(obj.contains("code") && obj.contains("message"))
+    if(!this->synonyms.isEmpty())
     {
-        QMessageBox::information(this,QString::number(obj.value("code").toDouble()),obj.value("message").toString());
-        return;
+        QStringList::const_iterator iter;
+        qDebug() << "Synonyms:";
+        for(iter = this->synonyms.constBegin(); iter != this->synonyms.constEnd(); ++iter)
+        {
+            qDebug() << *iter;
+        }
     }
 
-    // Set a detected language.
-    QString lang_code = obj.value("lang").toString();
-    ui->comboBoxSource->setCurrentText(languages[lang_code]);
-}
-
-void YandexTranslate::translate()
-{
-    QString source_lang = ui->comboBoxSource->currentText();
-    QString target_lang = ui->comboBoxTarget->currentText();
-
-    // Create a request.
-    QUrl url(QString("https://translate.yandex.net/api/v1.5/tr.json/translate?")
-             + QString("key=") + trnsl_api_key
-             + QString("&lang=")
-             + languages.key(source_lang) + "-" + languages.key(target_lang)
-             + QString("&text=") + ui->SourceEdit->toPlainText()
-             //+ QString("&format=") + <text format>
-             //+ QString("&options=") + <translation options>
-             );
-
-    QNetworkRequest request(url);
-    request.setHeader(QNetworkRequest::ContentTypeHeader, "application/x-www-form-urlencoded");
-
-    // Send the request.
-    QEventLoop loop;
-    QNetworkAccessManager manager;
-    connect(&manager, &QNetworkAccessManager::finished, &loop, &QEventLoop::quit);
-
-    // Receive a reply.
-    QNetworkReply* reply = manager.get(request);
-    loop.exec();
-
-    // Parse the reply.
-    QByteArray bytes = reply->readAll();
-    QJsonObject obj = QJsonDocument::fromJson(bytes).object();  // Keys are ("code", "lang", "text")
-
-    // Check the response codes.
-    if(obj.contains("code") && obj.contains("message"))
+    if(!this->meanings.isEmpty())
     {
-        QMessageBox::information(this,QString::number(obj.value("code").toDouble()),obj.value("message").toString());
-        return;
+        QStringList::const_iterator iter;
+        qDebug() << "Meanings:";
+        for(iter = this->meanings.constBegin(); iter != this->meanings.constEnd(); ++iter)
+        {
+            qDebug() << *iter;
+        }
     }
 
-    //QString dir = obj.value("lang").toString();   // Use the auto-defined translation direction from the server.
-
-    // Show the translated text.
-    QJsonArray text = obj.value("text").toArray();
-    ui->TargetEdit->clear();
-    for(QJsonValue v : text)
+    int i = 1;  // An example number.
+    QVector<Example>::iterator it;
+    for(it = this->examples.begin(); it != this->examples.end(); ++it, ++i)
     {
-        ui->TargetEdit->append(v.toString());
+        qDebug() << "Example" << i;
+        it->showExample();
     }
 }
 
-void YandexTranslate::getTranslationDirections()
+void DictionaryEntry::showEntry()
+{
+    QString header = (this->text + " [" + this->ts + "] " + this->pos); // Header of an entry.
+    qDebug() << header.trimmed();   // Get rid of spaces in the beginning and in the end of a string.
+
+    int i = 1;  // A translation number.
+    QVector<Translation>::iterator iter;
+    for(iter = this->translations.begin(); iter != this->translations.end(); ++iter, ++i)
+    {
+        qDebug() << "Translation" << i;
+        iter->showTranslation();
+    }
+}
+
+void YandexDictionary::getLanguages(QMap<QString, QString> languages_)
+{
+    this->languages = languages_;
+}
+
+void YandexDictionary::getTranslationDirections()
 {
     // Create a request.
     QUrl url(QString("https://dictionary.yandex.net/api/v1/dicservice.json/getLangs?")
@@ -198,7 +129,7 @@ void YandexTranslate::getTranslationDirections()
     // Parse the reply.
     QByteArray bytes = reply->readAll();
 
-    /* A different approach to parsing the reply:
+    /* Another approach to parsing the reply:
      * you can check a type of the reply data
      * in order to choose an appropriate strategy for the parsing process. */
     QJsonDocument doc = QJsonDocument::fromJson(bytes);
@@ -208,8 +139,10 @@ void YandexTranslate::getTranslationDirections()
         // Check the response codes.
         if(obj.contains("code") && obj.contains("message"))
         {
-            QMessageBox::information(this,QString::number(obj.value("code").toDouble()),obj.value("message").toString());
-            return;
+            QMessageBox messageBox;
+            messageBox.critical(0,QString::number(obj.value("code").toDouble()),obj.value("message").toString());
+            messageBox.setFixedSize(500,200);
+            // TODO: Error handling.
         }
     }
     else if(doc.isArray())
@@ -222,11 +155,8 @@ void YandexTranslate::getTranslationDirections()
     }
 }
 
-void YandexTranslate::getDictionaryEntry()
+QVector<DictionaryEntry> YandexDictionary::getDictionaryEntry(QString& source_lang, QString& target_lang, QString& text)
 {
-    QString source_lang = ui->comboBoxSource->currentText();
-    QString target_lang = ui->comboBoxTarget->currentText();
-
     QString dict_dir = languages.key(source_lang) + "-" + languages.key(target_lang);
 
     /* The following commented condition check is not necessary because
@@ -240,7 +170,7 @@ void YandexTranslate::getDictionaryEntry()
     QUrl url(QString("https://dictionary.yandex.net/api/v1/dicservice.json/lookup?")
              + QString("key=") + dict_api_key
              + QString("&lang=") + dict_dir
-             + QString("&text=") + ui->SourceEdit->toPlainText()
+             + QString("&text=") + text
              //+ QString("&ui=") + <the language of the user's interface for displaying the dictionary entry>
              //+ QString("&flags=") + <search options (bitmask of flags)>
              );
@@ -269,13 +199,15 @@ void YandexTranslate::getDictionaryEntry()
         /* By the way, it seems like Dictionary API never returns
          * ERR_OK code, i.e 200. This is strange because
          * the code is yet mentioned in the documentation. */
-        QMessageBox::information(this,QString::number(obj.value("code").toDouble()),obj.value("message").toString());
-        return;
+        QMessageBox messageBox;
+        messageBox.critical(0,QString::number(obj.value("code").toDouble()),obj.value("message").toString());
+        messageBox.setFixedSize(500,200);
+        // TODO: Error handling.
     }
 
     //QJsonObject head = obj.value("head").toObject();    // Result header (not used).
 
-    QVector<DictionaryEntry> entries;   // We will fill this QVector with data, i.e dictionary entries.
+    QVector<DictionaryEntry> entries;   // We will fill this QVector with data, i.e. dictionary entries.
     QJsonArray def = obj.value("def").toArray();    // Dictionary entries.
     for (QJsonValue v : def)
     {
@@ -370,10 +302,6 @@ void YandexTranslate::getDictionaryEntry()
         entries.push_back(entry);   // Add another entry.
     }
 
-    // We have collected all of our entries. Let's show them.
-    QVector<DictionaryEntry>::iterator iter;
-    for(iter = entries.begin(); iter != entries.end(); ++iter)
-    {
-        iter->showEntry();  // Look into dictionaryentry.cpp for the current implementation of showEntry().
-    }
+    // We have collected all of our entries. Let's return al of them.
+    return entries;
 }
