+++ dictionary_api.h
@@ -1,9 +1,12 @@
-#ifndef DICTIONARYENTRY_H
-#define DICTIONARYENTRY_H
+#ifndef DICTIONARY_API_H
+#define DICTIONARY_API_H
+
+#include <QObject>
 
 #include <QVector>
 #include <QString>
 #include <QStringList>
+#include <QMap>
 
 class Example
 {
@@ -66,4 +69,25 @@ private:
     QVector<Translation> translations; // Translations.
 };
 
-#endif // DICTIONARYENTRY_H
+class YandexDictionary : public QObject
+{
+    Q_OBJECT
+public:
+    YandexDictionary();
+    ~YandexDictionary();
+
+    QMap<QString, QString>& getLanguages() { return languages; }
+
+    void getTranslationDirections();// Get a list of translation directions from Dictionary API.
+    QVector<DictionaryEntry> getDictionaryEntry(QString& source_lang, QString& target_lang, QString& text);    // Get an automatically generated dictionary entry for a word or phrase.
+
+public slots:
+    void getLanguages(QMap<QString, QString> languages_);    // Get languages and their codes from Translate API.
+
+private:
+    QString dict_api_key = "";   // Put your Dictionary API key here.
+    QStringList directions; // Translation directions. They can be "en-en", "ru-ru", for example, but we don't consider this case.
+    QMap<QString, QString> languages; // Supported languages are values, their codes are keys. They're acquired from Translate API!
+};
+
+#endif // DICTIONARY_API_H
