+++ main.cpp
@@ -1,10 +1,10 @@
-#include "yandextranslate.h"
+#include "mainwindow.h"
 #include <QApplication>
 
 int main(int argc, char *argv[])
 {
     QApplication a(argc, argv);
-    YandexTranslate w;
+    MainWindow w;
     w.show();
 
     return a.exec();
